requirements.txt
numpy==1.24.3
pandas==2.0.2
matplotlib==3.7.1
scikit-learn==1.2.2
pillow==9.5.0
torch==2.0.1
torchvision==0.15.2
efficientnet-pytorch==0.7.1
opencv-python==4.7.0.72
face-recognition==1.3.0
mtcnn==0.1.1
fastapi==0.96.0
uvicorn==0.22.0
python-multipart==0.0.6
tqdm==4.65.0
pytest==7.3.1
gdown==4.7.1
requests==2.30.0